Dummy -> AlwaysEvaluative
